requirements.txt
# Bitget crypto scanner — minimal dependencies
requests>=2.31.0,<3
python-dotenv>=1.0.0,<2
